Per-line checksum (#21)
* Add persistent segtree

* Fix test

* Remove double semicolon

* Get rid of build()

* Add first version of the checksum script (#17)

When fed with a C++ code, outputs the code with a 3-letter hash
prepended to each line. TODO: implement the same functionality for
/export/build.

* Checksum script: Fix the handling of backslashes

* Export: prepare the TeX macros for per-line hashes

* Export: add per-line hashing to the PDF

* Ensure that all files end with a newline

Otherwise an automated testing of the check.sh script could end up with
false negatives (check.sh output stripped of the hashes could be
different from the actual file because check.sh appends the newline
automatically).

* Workaround in geo/geo.cpp because of our syntax highlighter

When there is a comment after a preprocessor directive, the highlighter
moves it to the next line. This didn't do anything bad, but now it
breaks line hashing.

* Line hashing: Simplify the regexes by omitting some edge cases

Don't handle */ comments at all, don't produce compileable code, just
delete whitespace.