Typesetting improvements (#26)
* [sc.sh modifications (merged into master since PR's creation)]

* Reduce the length of line hashes from 3 to 2.

The collision probability is 1/256, which is quite high, but multiplied
by the probability of a mistake per line, it gets low enough to be
ignore.

* defs.tex: split \setverb into two macros, adjust line-spacing constants

* defs.tex: fix whitespace issues in verbatim macros

* Decrease font sizes, use CM as the serif font, unify font-switching

To further increase the amount of code that fits in 25 pages, the font
sizes have once again been decreased.

In order for the yet-to-be inter-code texts to be consistent with inline
math, the serif font family has been changed to CMU Serif (TTF version
of the Computer Modern serif fonts).

Two new macros have been defined, \seriffam and \monofam. These macros
simplify the process of switching between the serif and mono fonts as
they redefine \rm, \bf and \it accordingly.

* Scale some of the maths fonts down

As stated in the comment, not all fonts were changed, only the most
used.

* Rename \calccharwidth to \calcchardimens, now it measures height as well

* Decrease page margins a lot

* Don't call \makeprintable globally since it is called locally by \verb

* Verbatim: use | as a delimiter, make inline verbatim more usable

If there were no delimiter, "\verb something\egroup" would print an
excess whitespace. This is a hotfix, there should be a more obvious way
of doing it.

Also, a macro \simpleverbatim has been defined, which is meant exactly
for the purposes of inline verbatim, since it doesn't mess with
inter-line spacing.